size_t causes compile error in cuda_stubs.cu

+++ cuda/cuda_stubs.cu
@@ -264,7 +264,7 @@ value ocaml_cuda_memcpy_device_to_symbol(value modulePtr,
   CUdeviceptr source = (CUdeviceptr)Int64_val(src);
 
   CUdeviceptr dest;
-  size_t bytes;
+  unsigned int bytes;
   CUresult result = cuModuleGetGlobal(&dest, &bytes, *cuModule, name);
 
   int offset = Int_val(ocaml_offset);
